refactor: move patients.json to assets/patients.json
